Navigation, AI, and entity system refactor & fixes
[General]:
- BSP plane key struct introduced for unique plane identification and debugging. (Not in-use yet)

[SVGame]:
- Major refactor of debug monster AI/navigation (svg_monster_testdummy_debug):
  - Introduced explicit state machine for AI, with states for idle, pursuit, trail, and sound investigation.
  - Navigation state and policy encapsulated in dedicated structs; all pathfinding and movement logic routed through new helpers.
  - Async navigation request handling hardened; pending requests cancelled/reset on state transitions.
  - Path-following and movement logic uses double precision and improved heuristics.
  - Breadcrumb and sound investigation logic improved; idle lookout now uses smooth yaw scanning.
  - Entity callbacks updated to use new state machine and navigation logic.
  - TU-local navigation helpers and linker stubs provided; save/restore descriptors updated.
  - Extensive comments and README updates for new workflow.
- Navigation system improvements and bug fixes:
  - Navmesh loading/saving includes generation statistics for validation; mismatches logged.
  - Navmesh file path construction centralized; loading returns status and path.
  - Agent bbox validation enforced at all boundaries.
  - Path-following APIs consistently convert feet-origin to nav-center; path process tracks start/goal and offset Z.
  - Async request queue supports "Preparing" status; robust cancellation and terminal state handling.
  - Edge rejection diagnostics improved; per-reason counters logged at multiple points.
  - Step/jump traversal uses correct drop caps and reacquire depths.
  - World-only traces for nav edge validation retry with Z lift to avoid precision artifacts.
  - Nav path policy uses physics constants; debug draw cvars tuned.
  - Legacy navmesh generation helpers removed.
- Entity and player state fields renamed for clarity (e.g., pain_debounce_time → debouncePainTime).
- Player trail API now uses svg_monster_testdummy_debug_t* for type safety.
- SVG_Util_IsEntityAudibleByPHS utility added for sound propagation checks.
- Trace/clip code uses explicit Vector3 bounds; octagon hull plane distance calculation updated.
- HQ2x/HQ4x upscaling code variable names clarified.
- Game DLL shutdown can be deferred for ERR_DROP; SV_Shutdown and SV_InitGameProgs updated.
- Cvar defaults for Vulkan validation and sun bounce range updated.
- Save descriptor macro ends with BAD field for safety.
- Monster testdummy header includes updated.
- Minor code style and comment improvements.

[SharedGame]:
- Monster testdummy navigation state members commented out; trail_time added for trail-following.

+++ inc/shared/math/qm_vector3.h
@@ -215,31 +215,31 @@ QM_API double QM_Vector3DotProductDP( const Vector3 &v1, const Vector3 &v2 ) {
 }
 
 // Calculate distance between two vectors
-QM_API float QM_Vector3Distance( const Vector3 &v1, const Vector3 &v2 ) {
-    const float dx = v2.x - v1.x;
-    const float dy = v2.y - v1.y;
-    const float dz = v2.z - v1.z;
-    const float result = sqrtf( dx * dx + dy * dy + dz * dz );
+QM_API double QM_Vector3Distance( const Vector3 &v1, const Vector3 &v2 ) {
+    const double dx = v2.x - v1.x;
+    const double dy = v2.y - v1.y;
+    const double dz = v2.z - v1.z;
+    const double result = std::sqrt( dx * dx + dy * dy + dz * dz );
 
     return result;
 }
 
 // Calculate square distance between two vectors
-QM_API float QM_Vector3DistanceSqr( const Vector3 &v1, const Vector3 &v2 ) {
-    const float dx = v2.x - v1.x;
-    const float dy = v2.y - v1.y;
-    const float dz = v2.z - v1.z;
-    const float result = dx * dx + dy * dy + dz * dz;
+QM_API double QM_Vector3DistanceSqr( const Vector3 &v1, const Vector3 &v2 ) {
+    const double dx = v2.x - v1.x;
+    const double dy = v2.y - v1.y;
+    const double dz = v2.z - v1.z;
+    const double result = dx * dx + dy * dy + dz * dz;
 
     return result;
 }
 
 // Calculate angle between two vectors
-QM_API float QM_Vector3Angle( const Vector3 &v1, const Vector3 &v2 ) {
+QM_API double QM_Vector3Angle( const Vector3 &v1, const Vector3 &v2 ) {
     const Vector3 cross = { v1.y * v2.z - v1.z * v2.y, v1.z * v2.x - v1.x * v2.z, v1.x * v2.y - v1.y * v2.x };
-    const float len = sqrtf( cross.x * cross.x + cross.y * cross.y + cross.z * cross.z );
-    const float dot = ( v1.x * v2.x + v1.y * v2.y + v1.z * v2.z );
-    const float result = atan2f( len, dot );
+    const double len = std::sqrt( cross.x * cross.x + cross.y * cross.y + cross.z * cross.z );
+    const double dot = ( v1.x * v2.x + v1.y * v2.y + v1.z * v2.z );
+    const double result = std::atan2( len, dot );
 
     return result;
 }
